PDJB-432: Give links their own file
